refactor: 관리자 페이지에 중복되는 코드 리팩토링
